Adds support for parentheticals in Tnex WHERE statements
Usage is query.and(clause => clause.where(...).orWhere(...)) etc.
